Add Comment Resolvers + Fix Relationship bugs.

+++ server/src/graphql/resolvers/topics.js
@@ -1,4 +1,6 @@
 import Topic from '../../models/Topic.js';
+import User from '../../models/User.js';
+import Article from '../../models/Article.js';
 
 const topicResolvers = {
 
@@ -60,6 +62,7 @@ const topicResolvers = {
                 });
 
                 await newTopic.save();
+                await User.findByIdAndUpdate(currentUser._id, { $push: { topics: newTopic._id } }, { new: true });
                 return newTopic;
             } 
             catch (err) {
@@ -103,13 +106,20 @@ const topicResolvers = {
 
                     throw new Error("You must be logged in to delete this topic.");
                 }
-                    
+
+                const topic = await Topic.findById(topicId);
+                if (!topic) {
+
+                    throw new Error("Article not found!");
+                }
+
                 if (currentUser.role !== "ADMIN") {
 
                     throw new Error("You don't have permission to delete this topic.");
                 }
 
                 const deletedTopic = await Topic.findByIdAndDelete(topicId);
+                await User.findByIdAndUpdate(topic.adminId, { $pull: { topics: topicId } }, { new: true });
                 return deletedTopic;
             } 
             catch (err) {
@@ -118,6 +128,36 @@ const topicResolvers = {
                 throw new Error(err.message || "Failed to delete topic!");
             }
         }
+    },
+
+    Topic: {
+
+        author: async (parent) => {
+
+            try {
+
+              const author = await User.findById(parent.authorId);
+              return author;
+            } 
+            catch (err) {
+                console.error(err);
+                throw new Error(err.message || "Failed to fetch topic's author!");
+            }
+        },
+
+        articles: async (parent) => {
+
+			try {
+
+				const articles = await Article.find({ topicId: parent._id });
+				return articles;
+			} 
+            catch (err) {
+                
+				console.error(err);
+				throw new Error(err.message || "Failed to fetch topic's articles!");
+			}
+		}
     }
 };
